ResourceServerAPI/src/test/java/responsabilitychains/CheckoutBased/handlers/Checkout/CheckoutAuthenticityCbasedHandlerTest.java
package responsabilitychains.CheckoutBased.handlers.Checkout;

import dev.ahernandez.checkoutresourceapi.logic.responsabilitychains.ChainResponse;
import dev.ahernandez.checkoutresourceapi.logic.responsabilitychains.checkoutbased.handlers.checkout.CheckoutAuthenticityCbasedHandler;
import dev.ahernandez.checkoutresourceapi.persistence.model.Checkout;
import dev.ahernandez.checkoutresourceapi.persistence.model.User;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContext;
import org.springframework.security.core.context.SecurityContextHolder;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.samePropertyValuesAs;

@DisplayName("CheckoutAuthenticityCHandler Test")
class CheckoutAuthenticityCbasedHandlerTest {

    @Test
    @DisplayName("Success CheckoutAuthenticityCHandler")
    void successHandler(){
        String userId = "";

        Authentication auth = Mockito.mock(Authentication.class);
        SecurityContext context = Mockito.mock(SecurityContext.class);
        Mockito.when(context.getAuthentication()).thenReturn(auth);
        SecurityContextHolder.setContext(context);

        Mockito.when(SecurityContextHolder.getContext().getAuthentication().getName())
                .thenReturn(userId);

        Checkout checkout = new Checkout();
        User user = new User();
        user.setUserId(userId);
        checkout.setUser(user);

        CheckoutAuthenticityCbasedHandler handler = new CheckoutAuthenticityCbasedHandler();

        ChainResponse actual = handler.handle(checkout);

        ChainResponse expected = new ChainResponse(true, "", checkout);

        assertThat(actual, samePropertyValuesAs(expected));
    }

    @Test
    @DisplayName("Fail CheckoutAuthenticityCHandler")
    void failHandler(){
        String userId = "";

        Authentication auth = Mockito.mock(Authentication.class);
        SecurityContext context = Mockito.mock(SecurityContext.class);
        Mockito.when(context.getAuthentication()).thenReturn(auth);
        SecurityContextHolder.setContext(context);

        Mockito.when(SecurityContextHolder.getContext().getAuthentication().getName())
                .thenReturn(userId);

        Checkout checkout = new Checkout();
        User user = new User();
        user.setUserId("diff");
        checkout.setUser(user);

        CheckoutAuthenticityCbasedHandler handler = new CheckoutAuthenticityCbasedHandler();

        ChainResponse actual = handler.handle(checkout);

        ChainResponse expected = new ChainResponse(false,"Checkout do not belongs to authenticated user");

        assertThat(actual, samePropertyValuesAs(expected));
    }
}
